--- app/assistant.config.ts
@@ -1,2 +1,0 @@
-// ID do assistente configurado
-export const assistantId = "asst_YjQgLU3twvivUChLp0SJ6Fbu";
