--- app/openai.ts
@@ -1,5 +1,0 @@
-import OpenAI from "openai";
-
-export const openai = new OpenAI({
-  apiKey: process.env.OPENAI_API_KEY,
-});
