+++ src/app/api/assistants/route.ts
@@ -1,4 +1,4 @@
-import { openai } from "@/app/openai";
+import { openai } from "@/server/utils/openai";
 
 export const runtime = "nodejs";
