+++ src/client/components/Modal/PrivacyModal/PrivacyModal.tsx
