+++ src/server/config/assistant.ts
@@ -1,0 +1,1 @@
+export const assistantId = process.env.OPENAI_ASSISTANT_ID || "asst_LTs4N9cLNB566TDDMV7ojIza";
