+++ src/shared/types/chat.ts
